+++ Tests/TgBusinessLogicTests/Using.cs
@@ -1,0 +1,12 @@
+// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
+
+global using Autofac;
+global using TgBusinessLogic.Contracts;
+global using TgBusinessLogic.Models;
+global using TgBusinessLogic.Services;
+global using TgBusinessLogicTests.Common;
+global using TgStorage.Contracts;
+global using TgStorage.Domain;
+global using TgStorage.Utils;
+global using TgStorageTests.Common;