+++ Tests/TgStorageTests/Using.cs
@@ -1,0 +1,33 @@
+// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
+
+global using Autofac;
+global using FluentAssertions;
+global using Microsoft.EntityFrameworkCore;
+global using Microsoft.Extensions.DependencyInjection;
+global using Microsoft.Extensions.Logging;
+global using NUnit.Framework;
+global using System.Diagnostics;
+global using TgBusinessLogic.Contracts;
+global using TgBusinessLogic.Helpers;
+global using TgBusinessLogic.Services;
+global using TgInfrastructure.Contracts;
+global using TgInfrastructure.Enums;
+global using TgInfrastructure.Helpers;
+global using TgStorage.Common;
+global using TgStorage.Contracts;
+global using TgStorage.Domain.Apps;
+global using TgStorage.Domain.Documents;
+global using TgStorage.Domain.Filters;
+global using TgStorage.Domain.Messages;
+global using TgStorage.Domain.Proxies;
+global using TgStorage.Domain.Sources;
+global using TgStorage.Domain.Stories;
+global using TgStorage.Domain.Users;
+global using TgStorage.Domain.Versions;
+global using TgStorage.Domain;
+global using TgStorage.Helpers;
+global using TgStorage.Repositories;
+global using TgStorage.Utils;
+global using TgStorageTests.Common;
+global using ZiggyCreatures.Caching.Fusion;