+++ Docs/CHANGELOG.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.7.160] - 2025-08-29
-### Added
-- Added hardware resource monitoring service
 ### Changed
 - Updated NuGet packages
+- Updated the `TgStorageTests` test project
+### Added
+- Added a hardware resource monitoring service
+- Added the `TgBusinessLogicTests` test project
 
 ## [0.7.140] - 2025-08-25
 ### Fixed